Add ProfilePage component with user profile management features
